feat: Rename component-shiki shikiComponent
